Fix bevel modifier bug #34611, limit bevel amount needed.
This is a quick fix that perhaps overestimates the point
of first geometry collision, but at least for now it should
allow models that used the old modifier and a too-big
bevel amount to not look awful.

The correct solution to this problem is much more involved
and I'll get to it later.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@55234 954f8c5b-7b00-dc11-b283-0030488c597c