Just added a comment:
To make circle select allow pass-through for view events, several issues have to be
tackled. 
 1) other modal ops run on top (border select), 
 2) middlemouse is used now 
 3) and what for tablet/trackpad or people without middlemouse?

The MMB deselection for border/circle is not optimal now... needs rethinking this.

Better would be to check on non-persistant-modality for circle, to start with gesture style
event like lasso or border can do now.



git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@55369 954f8c5b-7b00-dc11-b283-0030488c597c