Fix for 34703 Mix node (Hue, Saturation) update and rendering error Fix for 34494 Blender 2.65 regression test error - compo_map_uv_cubes.blend - stripe/artifact between cubes
Hue and saturation node has an early break when saturarion is 0. When this happened the input 1 color needed to be used. This behaviour was not merged.

When no FSAA is used in the ZCombine. a mask will be created, this mask will be antialiased and based on this mask the colors between the two images are blended. This was also behaviour that was not merged correctly. Now it is back making much better z-combines.

Hope nobody uses these gabs as a work around.

 - At Mind -
Jeroen & Monique



git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@55789 954f8c5b-7b00-dc11-b283-0030488c597c