Part II of the Freestyle branch merger: modifications to the trunk code base.
This commit concludes the two-part merger of the soc-2008-mxcurioni (Freestyle)
branch started in revision 55846.  A number of modifications and additions has
been made to the trunk code base to integrate the Freestyle non-photorealistic
(NPR) renderer into Blender.  All the changes in this commit were made by
applying a diff from the branch to the trunk (instead of svn merge, due to broken
svn:mergeinfo properties of the branch).

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@55847 954f8c5b-7b00-dc11-b283-0030488c597c