Fix for #35134.
The node link operator had a feature to automatically expose sockets in node groups when ctrl+shift+clicking on a node socket, which would create a node group input/output node. This was intended as a shortcut but conflicts with other features such as socket selection and viewer creation. It is also hardly necessary now that input/output nodes have an extension socket, which is much easier to use. Removed this expose functionality completely.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@56375 954f8c5b-7b00-dc11-b283-0030488c597c