Changes to image draw method options
It's now default to 2D textures, and no AUTO mode at this
moment, since detecting which method is the best not so
simple.

Image drawing could manually be switched to GLSL for tests
and feedback, but for default GLSL is not so much great.

Reason of this is huge images, where operations like panning
becomes dead slow comparing GLSL vs. 2D texture.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@56385 954f8c5b-7b00-dc11-b283-0030488c597c