Better API design for making text datablocks after loading.
An optional 'internal' argument was added to the bpy.data.texts.load() operator.  
The changes in revision 57153 were reverted, so that the is_in_memory and is_dirty
properties of text datablocks are not editable again.

In the C API layer, BKE_text_load_ex() was introduced to allow for optionally
making text datablocks internal after loading.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@57200 954f8c5b-7b00-dc11-b283-0030488c597c