Include DNA_scene_types before ED_object instead of forward enum declaration
Forward enum declaration is a bad idea, especially for C++ which requires
enum specification to dteermine which data type to use to store it.

Alternative would be to not use enum as an arument and pass it as int,
but actually would rather be strict on typing -- using explicit enum
as parameter type helps understanding the code and prevents possible
mistakes when using the function.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@57728 954f8c5b-7b00-dc11-b283-0030488c597c